🔨 [projects] Move conflict resolution from `services.import_` to `ProjectRepository.import_`
