+++ src/cutty/services/import_.py
@@ -1,9 +1,6 @@
 """Import changes from templates into projects."""
-import enum
 from pathlib import Path
 from typing import Optional
-
-import pygit2
 
 from cutty.projects.build import buildparentproject
 from cutty.projects.build import buildproject
@@ -11,29 +8,6 @@ from cutty.projects.config import ProjectConfig
 from cutty.projects.config import readprojectconfigfile
 from cutty.projects.messages import updatecommitmessage
 from cutty.projects.repository import ProjectRepository
-from cutty.util.git import MergeConflictError
-
-
-class Side(enum.Enum):
-    """The side of a conflict."""
-
-    ANCESTOR = 0
-    OURS = 1
-    THEIRS = 2
-
-
-def resolveconflicts(repositorypath: Path, path: Path, side: Side) -> None:
-    """Resolve conflicts in the given file."""
-    repository = pygit2.Repository(repositorypath)
-    pathstr = str(path.relative_to(repositorypath))
-    ancestor, ours, theirs = repository.index.conflicts[pathstr]
-    resolution = (ancestor, ours, theirs)[side.value]
-
-    del repository.index.conflicts[pathstr]
-
-    repository.index.add(resolution)
-    repository.index.write()
-    repository.checkout(strategy=pygit2.GIT_CHECKOUT_FORCE, paths=[pathstr])
 
 
 def import_(projectdir: Path, *, revision: Optional[str]) -> None:
@@ -70,18 +44,4 @@ def import_(projectdir: Path, *, revision: Optional[str]) -> None:
     # support, commits never have the same hash as their parents.
     assert commit != parent  # noqa: S101
 
-    try:
-        repository.import_(commit)
-    except MergeConflictError:
-        try:
-            resolveconflicts(projectdir, projectdir / "cutty.json", Side.THEIRS)
-        except KeyError:
-            pass
-
-        index = repository.project._repository.index
-        index.read()
-
-        if index.conflicts:
-            raise MergeConflictError.fromindex(index)
-
-        repository.continue_()
+    repository.import_(commit)
